// src/app/models/detalle-cita.ts
export interface DetalleCita {
	nombreCentroAten: string;
     nombreCompleto: string;
	 tipoDocumento: string;
	 documento: string;
	 telefono: string;
	 correoElectronico: string;
	 fechaCita: string;
	 horaCita: string;
	 especialidad: string;
	 profesional: string;
	 consultorio: string;
	 prestacion: string;
	 centroOperativo: string;
	 convenio: string;
	 usuarioAsignaCita: string;
	 fechaAsignacion: string;
	 codConvenio: string;
}